+++ src/Fantomas.Tests/CommentTests.fs
@@ -1843,3 +1843,26 @@ module Example =
     let dict2 =
         ConcurrentDictionary< (* some comment 2 *) int64, ConcurrentDictionary< (* some comment 3 *) int32, unit>>()
 """
+
+[<Test>]
+let ``correctly collect a double slash comment before a xml doc comment, 2152`` () =
+    formatSourceString
+        false
+        """
+// Maybe computation expression builder, copied from ExtCore library
+/// https://github.com/jack-pappas/ExtCore/blob/master/ExtCore/Control.fs
+[<Sealed>]
+type MaybeBuilder() = class end
+"""
+        config
+    |> prepend newline
+    |> should
+        equal
+        """
+// Maybe computation expression builder, copied from ExtCore library
+/// https://github.com/jack-pappas/ExtCore/blob/master/ExtCore/Control.fs
+[<Sealed>]
+type MaybeBuilder() =
+    class
+    end
+"""